+++ hs-bindgen/test-artefacts/fixtures/macros/macro_type_ptr_qualifiers.2.raw/Example.hs
@@ -8,14 +8,18 @@ module Example
     )
   where
 
+import qualified HsBindgen.Runtime.Macro as Macro
+import qualified HsBindgen.Runtime.Support as BG
+
 {-| __C declaration:__ @macro PtrToVoid@
 
     __defined at:__ @macros\/macro_type_ptr_qualifiers.h 2:9@
 
     __exported by:__ @macros\/macro_type_ptr_qualifiers.h@
 -}
-ptrToVoid :: [String]
-ptrToVoid = ["void", "*"]
+ptrToVoid :: Macro.Raw BG.Text
+ptrToVoid =
+  Macro.objectLike "PtrToVoid" ["void", "*"]
 
 {-| __C declaration:__ @macro PtrToConstVoidL@
 
@@ -23,8 +27,9 @@ ptrToVoid = ["void", "*"]
 
     __exported by:__ @macros\/macro_type_ptr_qualifiers.h@
 -}
-ptrToConstVoidL :: [String]
-ptrToConstVoidL = ["const", "void", "*"]
+ptrToConstVoidL :: Macro.Raw BG.Text
+ptrToConstVoidL =
+  Macro.objectLike "PtrToConstVoidL" ["const", "void", "*"]
 
 {-| __C declaration:__ @macro PtrToConstVoidR@
 
@@ -32,8 +37,9 @@ ptrToConstVoidL = ["const", "void", "*"]
 
     __exported by:__ @macros\/macro_type_ptr_qualifiers.h@
 -}
-ptrToConstVoidR :: [String]
-ptrToConstVoidR = ["void", "const", "*"]
+ptrToConstVoidR :: Macro.Raw BG.Text
+ptrToConstVoidR =
+  Macro.objectLike "PtrToConstVoidR" ["void", "const", "*"]
 
 {-| __C declaration:__ @macro PtrToConstIntL@
 
@@ -41,8 +47,9 @@ ptrToConstVoidR = ["void", "const", "*"]
 
     __exported by:__ @macros\/macro_type_ptr_qualifiers.h@
 -}
-ptrToConstIntL :: [String]
-ptrToConstIntL = ["const", "int", "*"]
+ptrToConstIntL :: Macro.Raw BG.Text
+ptrToConstIntL =
+  Macro.objectLike "PtrToConstIntL" ["const", "int", "*"]
 
 {-| __C declaration:__ @macro PtrToConstIntR@
 
@@ -50,8 +57,9 @@ ptrToConstIntL = ["const", "int", "*"]
 
     __exported by:__ @macros\/macro_type_ptr_qualifiers.h@
 -}
-ptrToConstIntR :: [String]
-ptrToConstIntR = ["int", "const", "*"]
+ptrToConstIntR :: Macro.Raw BG.Text
+ptrToConstIntR =
+  Macro.objectLike "PtrToConstIntR" ["int", "const", "*"]
 
 {-| __C declaration:__ @macro ConstPtrToInt@
 
@@ -59,5 +67,6 @@ ptrToConstIntR = ["int", "const", "*"]
 
     __exported by:__ @macros\/macro_type_ptr_qualifiers.h@
 -}
-constPtrToInt :: [String]
-constPtrToInt = ["int", "*", "const"]
+constPtrToInt :: Macro.Raw BG.Text
+constPtrToInt =
+  Macro.objectLike "ConstPtrToInt" ["int", "*", "const"]