+++ hs-bindgen/test-artefacts/fixtures/macros/redeclaration/identical_syntax.2.raw/Example.hs
@@ -1,0 +1,16 @@
+module Example
+    ( Example.t
+    )
+  where
+
+import qualified HsBindgen.Runtime.Macro as Macro
+import qualified HsBindgen.Runtime.Support as BG
+
+{-| __C declaration:__ @macro T@
+
+    __defined at:__ @macros\/redeclaration\/identical_syntax.h 5:9@
+
+    __exported by:__ @macros\/redeclaration\/identical_syntax.h@
+-}
+t :: Macro.Raw BG.Text
+t = Macro.objectLike "T" ["A"]